@@ -88,12 +91,24 @@
   "deprecated": false,
   "description": "Node's domain module for the web browser. This is merely an evented try...catch with the same API as node, nothing more.",
   "devDependencies": {
-    "assert-helpers": "^4.1.0",
-    "eslint": "^1.10.3",
-    "joe": "^1.6.0",
-    "joe-reporter-console": "^1.2.1",
-    "projectz": "^1.0.8"
+    "assert-helpers": "^4.5.0",
+    "eslint": "^4.16.0",
+    "joe": "^2.0.2",
+    "joe-reporter-console": "^2.0.1",
+    "projectz": "^1.4.0"
   },
+  "editions": [
+    {
+      "description": "Source + ES5 + Require",
+      "directory": "source",
+      "entry": "index.js",
+      "syntaxes": [
+        "javascript",
+        "es5",
+        "require"
+      ]
+    }
+  ],
   "engines": {
     "node": ">=0.4",
     "npm": ">=1.2"
