@@ -1,9 +1,20 @@
-
+//
 module.exports = {
-    entry: './src/main.js',
-    output: {
-        path: './dist/js',
-        publicPath: '/dist/',
-        filename: 'bundle.js'
+//     module: {
+//         loaders: [
+//             {
+//                 test: /\.js$/,
+//                 exclude: /(node_modules|bower_components)/,
+//                 use: {
+//                     loader: 'babel-loader',
+//                     options: {
+//                         presets: ['@babel/preset-env']
+//                     }
+//                 }
+//             }
+//         ]
+//     }
+    node: {
+        fs: 'empty'
     }
 };
